project changes

+++ src/main/java/com/tts/ecommerce/controller/CheckoutController.java
@@ -9,7 +9,8 @@ import org.springframework.web.bind.annotation.RequestParam;
 import com.tts.ecommerce.model.ChargeRequest;
 
 @Controller
-public class CheckoutController {
+public class CheckoutController 
+{
 
 	@Value("${STRIPE_PUBLIC_KEY}")
     private String stripePublicKey;
@@ -17,7 +18,7 @@ public class CheckoutController {
     @RequestMapping("/checkout")
     public String checkout(@RequestParam("amount") double amount, Model model) {
     	amount *= 100; 
-        model.addAttribute("amount", (double) amount); // in cents
+        model.addAttribute("amount", (int) amount); // in cents
         model.addAttribute("stripePublicKey", stripePublicKey);
         model.addAttribute("currency", ChargeRequest.Currency.USD);
         return "checkout";
